Use named classes for derived ArgBuilders
